test(demo): adding a emo of testing style
Here just we demo on DOM selection and promises

re #50
