+++ .nyc_output/processinfo/0ef8d3d1-e337-4ad9-b17d-e29ca4d64c1c.json
@@ -1,0 +1,1 @@
+{"parent":null,"pid":10102,"argv":["/Users/julien.froidefond/.nvm/versions/node/v12.15.0/bin/node","/Users/julien.froidefond/Sites/strapi-bookmarks/node_modules/.bin/mocha","-r","esm"],"execArgv":[],"cwd":"/Users/julien.froidefond/Sites/strapi-bookmarks","time":1617258234972,"ppid":10095,"coverageFilename":"/Users/julien.froidefond/Sites/strapi-bookmarks/.nyc_output/0ef8d3d1-e337-4ad9-b17d-e29ca4d64c1c.json","externalId":"","uuid":"0ef8d3d1-e337-4ad9-b17d-e29ca4d64c1c","files":[]}
