+++ .nyc_output/processinfo/7845cc95-15dd-4aba-b1cc-6c9d5c1c9924.json
@@ -1,0 +1,1 @@
+{"parent":"0ef8d3d1-e337-4ad9-b17d-e29ca4d64c1c","pid":10103,"argv":["/Users/julien.froidefond/.nvm/versions/node/v12.15.0/bin/node","/Users/julien.froidefond/Sites/strapi-bookmarks/node_modules/mocha/lib/cli/cli.js","--no-config","--no-package","--diff","--extension","js","--extension","cjs","--extension","mjs","--reporter","spec","--slow","75","--timeout","2000","--ui","bdd","--watch-ignore","node_modules","--watch-ignore",".git"],"execArgv":["--require","esm"],"cwd":"/Users/julien.froidefond/Sites/strapi-bookmarks","time":1617258235256,"ppid":10102,"coverageFilename":"/Users/julien.froidefond/Sites/strapi-bookmarks/.nyc_output/7845cc95-15dd-4aba-b1cc-6c9d5c1c9924.json","externalId":"","uuid":"7845cc95-15dd-4aba-b1cc-6c9d5c1c9924","files":[]}
