+++ package.json
@@ -1,16 +1,22 @@
 {
   "devDependencies": {
+    "chai": "^4.3.4",
     "clean-webpack-plugin": "3.0.0",
     "copy-webpack-plugin": "8.1.0",
     "css-loader": "5.2.0",
     "cz-conventional-changelog": "3.3.0",
+    "esm": "^3.2.25",
     "file-loader": "6.2.0",
     "fs-extra": "9.1.0",
     "html-loader": "2.1.2",
     "html-webpack-plugin": "5.3.1",
     "husky": "6.0.0",
+    "jsdom": "^16.5.2",
+    "mocha": "^8.3.2",
+    "nyc": "^15.1.0",
     "prettier": "2.2.1",
     "pretty-quick": "3.1.0",
+    "sinon-chrome": "^3.0.1",
     "style-loader": "2.0.0",
     "webpack": "5.28.0",
     "webpack-cli": "4.6.0",
@@ -20,7 +26,8 @@
   "scripts": {
     "build": "node utils/build.js",
     "start": "node utils/webserver.js",
-    "commit": "cz"
+    "commit": "cz",
+    "test": "nyc mocha -r esm"
   },
   "husky": {
     "hooks": {
