+++ src/popup.html
@@ -1,6 +1,7 @@
 <!DOCTYPE html>
 <head>
   <meta http-equiv="Content-Type" content="text/html; charset=UTF-8" />
+  <link rel="stylesheet" href="./css/popup.css" />
 </head>
 <body>
   <div id="title">
